backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
cryptography==41.0.7

# Email
aiosmtplib==3.0.1
email-validator==2.1.0
emails==0.6
jinja2==3.1.2

# File handling
aiofiles==23.2.1
python-magic==0.4.27
pillow==10.1.0

# Rate limiting and security
slowapi==0.1.9
limits==3.6.0
redis==5.0.1
pyotp==2.9.0

# Environment and configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.0.3

# Date/time handling
python-dateutil==2.8.2

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0